backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gitpython==3.1.40
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.25.2
python-dotenv==1.0.0
networkx==3.2.1
tree-sitter==0.20.4
aiofiles==23.2.1
python-multipart==0.0.6
jinja2==3.1.2
openai==1.6.1
